requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
ultralytics==8.0.114
ultralytics-thop==2.0.0
cvzone==1.6.1
easyocr==1.7.2
pymongo==4.8.0
python-dotenv==1.0.1
olamaps==0.5.2